reorganized components folder
