+++ src/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import Planner from "./components/Planner.vue"
-import Navbar from "./components/Navbar.vue"
+import Planner from "@/components/Planner.vue"
+import Navbar from "@/components/Navbar.vue"
 </script>
 
 <template>
