+++ src/components/akplanner/CraftButton.vue
