+++ src/components/akplanner/ItemDisplayCell.vue
@@ -5,8 +5,8 @@ import { stages } from '@/data/farmingdata';
 import { Inventory } from '@/store/store-inventory-functions';
 import { storeToRefs } from 'pinia';
 import { usePlannerStore } from '@/store/planner-store';
-import CraftButton from '@/components/CraftButton.vue';
-import ImageFinder from '@/components/ImageFinder.vue';
+import CraftButton from '@/components/akplanner/CraftButton.vue';
+import ImageFinder from '@/components/akplanner/ImageFinder.vue';
 
 export interface Props {
     item: Item;
